requirements.txt
python-telegram-bot 
requests
instagrapi
moviepy
pillow
imageio
imageio[ffmpeg]
numpy
tqdm